Update refcount survey to track data in-memory.
This is significantly faster and somewhat easier than dumping the
trace (in fact, the slowdown isn't terribly noticeable anymore).

However, the results will now be significantly less accurate,
since we keep track of every increment for an object, rather
than just the highest value it achieves. Doing so would be
much more difficult since we'd need to track the size of every
live object.